Add custom Valet driver for simpler local dev
The built-in PHP simple server is bug-ridden and doesn’t do a great job
of emulating an Apache/nginx environment, but the Valet/Caddy
environment is pretty friendly. Initially used to get around the
(incredible) simple-server limitation with dots in URLs. This was
important, as Kirby’s Content Representations make use of “extensions”
that don’t resolve to actual files, and therefore break when running on
the built-in server.